requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
abseil-cpp=20210324.2=hbdafb3b_0
absl-py=1.1.0=pypi_0
aiohttp=3.8.1=py39hb18efdd_1
aiosignal=1.2.0=pyhd8ed1ab_0
appnope=0.1.3=pyhd8ed1ab_0
argon2-cffi=21.3.0=pyhd8ed1ab_0
argon2-cffi-bindings=21.2.0=py39hb18efdd_2
asttokens=2.0.5=pyhd8ed1ab_0
astunparse=1.6.3=pyhd8ed1ab_0
async-timeout=4.0.2=pyhd8ed1ab_0
attrs=21.4.0=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
beautifulsoup4=4.11.1=pyha770c72_0
bleach=5.0.0=pyhd8ed1ab_0
blinker=1.4=py_1
brotli=1.0.9=h1c322ee_7
brotli-bin=1.0.9=h1c322ee_7
brotlipy=0.7.0=py39hb18efdd_1004
bzip2=1.0.8=h3422bc3_4
c-ares=1.18.1=h3422bc3_0
ca-certificates=2022.6.15=h4653dfc_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cachetools=5.0.0=pyhd8ed1ab_0
certifi=2022.6.15=py39h2804cbe_0
cffi=1.15.0=py39h52b1de0_0
charset-normalizer=2.0.12=pyhd8ed1ab_0
click=8.1.3=py39h2804cbe_0
colorama=0.4.5=pyhd8ed1ab_0
cryptography=37.0.2=py39h4cb6239_0
cycler=0.11.0=pyhd8ed1ab_0
dataclasses=0.8=pyhc8e2a94_3
debugpy=1.6.0=py39h0ef5a74_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
dill=0.3.5.1=pyhd8ed1ab_0
docker-pycreds=0.4.0=pypi_0
entrypoints=0.4=pyhd8ed1ab_0
executing=0.8.3=pyhd8ed1ab_0
fastcore=1.4.5=pypi_0
flatbuffers=1.12=pypi_0
flit-core=3.7.1=pyhd8ed1ab_0
fonttools=4.33.3=py39h9eb174b_0
freetype=2.10.4=h17b34a0_1
frozenlist=1.3.0=py39hb18efdd_1
future=0.18.2=py39h2804cbe_5
gast=0.4.0=pypi_0
giflib=5.2.1=h27ca646_2
gitdb=4.0.9=pypi_0
gitpython=3.1.27=pypi_0
google-auth=2.8.0=pyh6c4a22f_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
google-pasta=0.2.0=pyh8c360ce_0
googleapis-common-protos=1.56.3=py39h2804cbe_0
grpc-cpp=1.45.2=h42cd831_3
grpcio=1.45.0=py39hcf421d0_0
h5py=3.6.0=nompi_py39hd982b79_100
hdf5=1.12.1=nompi_hf9525e8_104
icu=70.1=h6b3803e_0
idna=3.3=pyhd8ed1ab_0
importlib-metadata=4.11.4=py39h2804cbe_0
importlib_resources=5.8.0=pyhd8ed1ab_0
ipykernel=6.15.0=pyh736e0ef_0
ipython=8.4.0=py39h2804cbe_0
ipython_genutils=0.2.0=py_1
ipywidgets=7.7.0=pyhd8ed1ab_0
jedi=0.18.1=py39h2804cbe_1
jinja2=3.1.2=pyhd8ed1ab_1
joblib=1.1.0=pyhd8ed1ab_0
jpeg=9e=h1c322ee_1
jsons=1.6.3=pypi_0
jsonschema=4.6.0=pyhd8ed1ab_0
jupyter=1.0.0=py39h2804cbe_7
jupyter_client=7.3.4=pyhd8ed1ab_0
jupyter_console=6.4.3=pyhd8ed1ab_0
jupyter_core=4.10.0=py39h2804cbe_0
jupyterlab_pygments=0.2.2=pyhd8ed1ab_0
jupyterlab_widgets=1.1.0=pyhd8ed1ab_0
keras=2.9.0=pypi_0
keras-preprocessing=1.1.2=pyhd8ed1ab_0
kiwisolver=1.4.3=py39hab5e169_0
krb5=1.19.3=hf9b2bbe_0
lcms2=2.12=had6a04f_0
lerc=3.0=hbdafb3b_0
libblas=3.9.0=15_osxarm64_openblas
libbrotlicommon=1.0.9=h1c322ee_7
libbrotlidec=1.0.9=h1c322ee_7
libbrotlienc=1.0.9=h1c322ee_7
libcblas=3.9.0=15_osxarm64_openblas
libclang=14.0.1=pypi_0
libcurl=7.83.1=h2fcd78c_0
libcxx=14.0.5=h04bba0f_1
libdeflate=1.12=he4db4b2_0
libedit=3.1.20191231=hc8eb9b7_2
libev=4.33=h642e427_1
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0.dev0=11_0_1_hf114ba7_23
libgfortran5=11.0.1.dev0=hf114ba7_23
libiconv=1.16=h642e427_0
liblapack=3.9.0=15_osxarm64_openblas
libnghttp2=1.47.0=he723fca_0
libopenblas=0.3.20=openmp_h2209c59_0
libpng=1.6.37=hf7e6567_2
libprotobuf=3.20.1=h332123e_0
libsodium=1.0.18=h27ca646_1
libssh2=1.10.0=hb80f160_2
libtiff=4.4.0=hcbbed22_1
libwebp=1.2.2=h0d20362_0
libwebp-base=1.2.2=h3422bc3_1
libxcb=1.13=h9b22ae9_1004
libxml2=2.9.14=h035c1df_0
libxslt=1.1.35=ha8fb468_0
libzlib=1.2.12=ha287fd2_1
llvm-openmp=14.0.4=hd125106_0
lxml=4.9.0=py39h9eb174b_0
lz4-c=1.9.3=hbdafb3b_1
markdown=3.3.7=pyhd8ed1ab_0
markupsafe=2.1.1=py39hb18efdd_1
matplotlib=3.5.2=py39hdf13c20_0
matplotlib-base=3.5.2=py39h4ee150f_0
matplotlib-inline=0.1.3=pyhd8ed1ab_0
mistune=0.8.4=py39h5161555_1005
multidict=6.0.2=py39hb18efdd_1
munkres=1.1.4=pyh9f0ad1d_0
nbclient=0.5.13=pyhd8ed1ab_0
nbconvert=6.4.5=py39h2804cbe_0
nbformat=5.4.0=pyhd8ed1ab_0
ncurses=6.3=h07bb92c_1
nest-asyncio=1.5.5=pyhd8ed1ab_0
notebook=6.4.12=pyha770c72_0
numpy=1.22.4=py39h7df2422_0
oauthlib=3.2.0=pyhd8ed1ab_0
openjpeg=2.4.0=h062765e_1
openssl=1.1.1p=ha287fd2_0
opt_einsum=3.3.0=pyhd8ed1ab_1
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.2=py39hd2dba81_2
pandas-datareader=0.10.0=pyh6c4a22f_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
pathtools=0.1.2=pypi_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=9.1.1=py39h1b8be2f_1
pip=22.1.2=pyhd8ed1ab_0
prometheus_client=0.14.1=pyhd8ed1ab_0
promise=2.3=py39h2804cbe_6
prompt-toolkit=3.0.29=pyha770c72_0
prompt_toolkit=3.0.29=hd8ed1ab_0
protobuf=3.19.4=pypi_0
psutil=5.9.1=py39h9eb174b_0
pthread-stubs=0.4=h27ca646_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.12.0=pyhd8ed1ab_0
pyjwt=2.4.0=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyrsistent=0.18.1=py39hb18efdd_1
pysocks=1.7.1=py39h2804cbe_5
python=3.9.13=hc596b02_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-dotenv=0.20.0=pypi_0
python-fastjsonschema=2.15.3=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytz=2022.1=pyhd8ed1ab_0
pyu2f=0.1.5=pyhd8ed1ab_0
pyyaml=6.0=py39hb18efdd_4
pyzmq=23.1.0=py39h8faa4b9_0
re2=2022.04.01=h6b3803e_0
readline=8.1.2=h46ed386_0
requests=2.28.0=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.8=pyhd8ed1ab_0
scikit-learn=1.1.1=py39h255bef5_0
scipy=1.8.1=py39h14896cb_0
send2trash=1.8.0=pyhd8ed1ab_0
sentry-sdk=1.6.0=pypi_0
setproctitle=1.2.3=pypi_0
setuptools=62.6.0=py39h2804cbe_0
shortuuid=1.0.9=pypi_0
six=1.15.0=pyh9f0ad1d_0
smmap=5.0.0=pypi_0
snappy=1.1.9=h39c3846_1
soupsieve=2.3.1=pyhd8ed1ab_0
sqlite=3.38.5=h40dfcc0_0
stack_data=0.3.0=pyhd8ed1ab_0
tensorboard=2.9.1=pypi_0
tensorboard-data-server=0.6.0=py39hbe5e4b8_2
tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
tensorflow=2.8.1=cpu_py39h2839aeb_0
tensorflow-base=2.8.1=cpu_py39h84f6f83_0
tensorflow-datasets=4.3.0=pyhd8ed1ab_0
tensorflow-deps=2.9.0=0
tensorflow-estimator=2.9.0=pypi_0
tensorflow-macos=2.9.2=pypi_0
tensorflow-metadata=1.2.0=pyhd8ed1ab_0
tensorflow-metal=0.5.0=pypi_0
termcolor=1.1.0=py_2
terminado=0.15.0=py39h2804cbe_0
testpath=0.6.0=pyhd8ed1ab_0
threadpoolctl=3.1.0=pyh8a188c0_0
tk=8.6.12=he1e0b03_0
tornado=6.1=py39hb18efdd_3
tqdm=4.64.0=pyhd8ed1ab_0
traitlets=5.3.0=pyhd8ed1ab_0
tweepy=4.10.0=pypi_0
typing-extensions=4.2.0=hd8ed1ab_1
typing_extensions=4.2.0=pyha770c72_1
typish=1.9.3=pypi_0
tzdata=2022a=h191b570_0
unicodedata2=14.0.0=py39hb18efdd_1
urllib3=1.26.9=pyhd8ed1ab_0
wandb=0.12.19=pypi_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
werkzeug=2.1.2=pyhd8ed1ab_1
wheel=0.37.1=pyhd8ed1ab_0
widgetsnbextension=3.6.0=pyha770c72_1
wrapt=1.14.1=py39h9eb174b_0
xorg-libxau=1.0.9=h27ca646_0
xorg-libxdmcp=1.1.3=h27ca646_0
xz=5.2.5=h642e427_1
yaml=0.2.5=h3422bc3_2
yarl=1.7.2=py39hb18efdd_2
zeromq=4.3.4=hbdafb3b_1
zipp=3.8.0=pyhd8ed1ab_0
zlib=1.2.12=ha287fd2_1
zstd=1.5.2=hd705a24_1
